correção delete

+++ resources/views/pages/fornecedores/index.blade.php
@@ -42,7 +42,7 @@
                                     </a>
 
                                     <meta name='csrf-token' content="{{ csrf_token() }}"/>
-                                    <button onclick="deleteSupply({{$supply->id}})" class="btn btn-danger btn-sm">
+                                    <button onclick="deleteSupply('{{$supply->id}}', '{{$supply->name}}')" class="btn btn-danger btn-sm">
                                         Excluir
                                     </button>
                                 </td>
@@ -58,21 +58,44 @@
 <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
 
 <script>
-    function deleteSupply(id) {
-        axios.delete('{{route("supply.delete")}}', {
-            data: {
-                idForne: id,
+    function deleteSupply(id, name) {
+        Swal.fire({
+            title: 'Atenção',
+            html: `Deseja realmente excluir o fornecedor <b>"${name}"</b>?`,
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#868e96',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sim, remover!',
+            cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete('{{route("supply.delete")}}', {
+                    data: {
+                        idForne: id,
+                    }
+                })
+                .then(response => {
+
+                    if(response.data.success) {
+                        Swal.fire({
+                            title: 'Excluido com sucesso',
+                            icon: 'success'
+                        })
+                        .then(res => {
+                            document.location.reload();
+                        })
+                    }
+                })
+                .catch(error => {
+                    Swal.fire(
+                        'Oops',
+                        'Não foi possivel excluir!!',
+                        'error'
+                    )
+                    
+                });
             }
-        })
-        .then(function(response) {
-            //verificar se deu succes igual a true 
-            //e fazer um reload da página
-            console.log(response);
-        })
-        .catch(function(error){
-            // exibir um alert para mostrar o erro
-            console.log(error);
-            
         });
     }
 </script>
